base/requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
blas=1.0=mkl
ca-certificates=2019.5.15=1
certifi=2019.6.16=py36_1
cffi=1.12.3=py36h2e261b9_0
cudatoolkit=9.2=0
cycler=0.10.0=py_1
fontconfig=2.13.1=h86ecdb6_1001
freetype=2.9.1=h8a8886c_1
icu=64.2=he1b5a44_1
intel-openmp=2019.4=243
joblib=0.13.2=py36_0
jpeg=9b=h024ee3a_2
kiwisolver=1.1.0=py36hc9558a2_0
libblas=3.8.0=12_openblas
libcblas=3.8.0=12_openblas
libedit=3.1.20181209=hc058e9b_0
libffi=3.2.1=hd88cf55_4
libgcc-ng=9.1.0=hdf63c60_0
libgfortran-ng=7.3.0=hdf63c60_0
libiconv=1.15=h516909a_1005
liblapack=3.8.0=12_openblas
libopenblas=0.3.7=h6e990d7_1
libpng=1.6.37=hbc83047_0
libstdcxx-ng=9.1.0=hdf63c60_0
libtiff=4.0.10=h2733197_2
libuuid=2.32.1=h14c3975_1000
libxml2=2.9.9=hee79883_5
matplotlib=3.1.1=py36_1
matplotlib-base=3.1.1=py36he7580a8_1
mkl=2019.4=243
mkl-service=2.0.2=py36h7b6447c_0
ncurses=6.1=he6710b0_1
ninja=1.9.0=py36hfd86e86_0
numpy=1.17.2=py36h95a1406_0
olefile=0.46=py36_0
openssl=1.1.1c=h7b6447c_1
pandas=0.25.1=py36he6710b0_0
pillow=6.1.0=py36h34e0f95_0
pip=19.2.2=py36_0
pycparser=2.19=py36_0
pyparsing=2.4.2=py_0
pyqt=4.11.4=py36_3
python=3.6.9=h265db76_0
python-dateutil=2.8.0=py_0
pytorch=1.2.0=py3.6_cuda9.2.148_cudnn7.6.2_0
pytz=2019.2=py_0
pyyaml=5.1.2=py36h7b6447c_0
qt=4.8.7=2
readline=7.0=h7b6447c_5
scikit-learn=0.21.2=py36hd81dba3_0
scipy=1.3.1=py36h7c811a0_0
setuptools=41.0.1=py36_0
sip=4.18=py36_1
six=1.12.0=py36_0
sqlite=3.29.0=h7b6447c_0
tk=8.6.9=hed695b0_1002
torchvision=0.4.0=py36_cu92
tornado=6.0.3=py36h516909a_0
tqdm=4.32.1=py_0
wheel=0.33.4=py36_0
xz=5.2.4=h14c3975_4
yaml=0.1.7=had09818_2
zlib=1.2.11=h7b6447c_3
zstd=1.3.7=h0b5b093_0